Loader::Find()
Added a method to the loader which can find all sub-classes /
implementations of a class / interface. The method must be called with
at least one argument (the name of the class / interface to find
subclasses / implementations of).

By default Loader::Find() will look for 'extra' classes. However, it is
possible to specify something else (e.g. 'ctrl' for controllers) as the
second argument.

Finally, by default, only packages which have already been loaded when
the method is called are considered. Passing false as the 3rd argument
will cause packages to be loaded.